Big update
No new features. Just preparing a configure sheet ideally for cpu and
player settings and stuff but I don’t even know how to create an
NSWindow NSObject to do that yet. Gonna learn for you (me. unless
you’re not me. in which case, hello there! please leave a comment.)

This update fixes the super-obvious glitchy lag that looks like someone
created it on purpose. The problem was that I was only allowing one key
to be pressed every time it checked. Instead, it now updates whether
keys are pressed every time it checks, and that changes whether the
paddles are moving.

It was really simple and easy. Just ⌘-X and ⌘-V (or ⌘-C for C-marangu)
also Unicode Rocks
{ 🅒Ⓜ️ⒶⓇⒶⓃⒼⓊ }